requirements.txt
fastapi
uvicorn[standard]
filelock
pydantic
python-multipart
groq
openai
python-dotenv
